apex_ai_engine/mcp_server/requirements.txt
# APEX AI MCP SERVER REQUIREMENTS
# ===============================
# Python dependencies for the Model Context Protocol server

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Data Processing
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.3
pandas==2.0.3

# Configuration Management
pyyaml==6.0.1
python-dotenv==1.0.0

# Database Integration
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI/ML Dependencies
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196

# Face Recognition
face-recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.0.1

# Audio Processing
pydub==0.25.1
soundfile==0.12.1

# HTTP Client
aiohttp==3.9.0
httpx==0.25.2

# OpenAI Integration
openai==1.3.7

# Azure Cognitive Services
azure-cognitiveservices-speech==1.32.1

# Async Utilities
asyncio-mqtt==0.13.0

# Security
cryptography==41.0.8
bcrypt==4.1.1

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
